Cut down dependencies
