+++ Cargo.toml
@@ -59,11 +59,10 @@ winauth = { version = "0.0.4", optional = true }
 [target.'cfg(unix)'.dependencies]
 libgssapi = { version = "0.4.5", optional = true, default-features = false }
 
-[target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
-opentls = { version = "0.2.1", features = ["io-async-std"]}
-
-[target.'cfg(not(any(target_os = "macos", target_os = "ios")))'.dependencies]
-async-native-tls = { version = "0.4", features = ["runtime-async-std"] }
+[dependencies.async-native-tls]
+version = "0.4"
+features = ["runtime-async-std"]
+optional = true
 
 [dependencies.tokio]
 version = "1.0"
@@ -119,21 +118,17 @@ optional = true
 version = "1.12.0"
 optional = true
 
-[dependencies.rustls]
-version = "0.20.4"
-features = ["dangerous_configuration"]
-optional = true
-
 [dependencies.tokio-rustls]
 version = "0.23.3"
 optional = true
-
-[dependencies.hyper-rustls]
-version = "0.23.0"
-optional = true
+features = ["dangerous_configuration"]
 
 [dependencies.rustls-pemfile]
-version = "0.3.0"
+version = "0.3"
+optional = true
+
+[dependencies.rustls-native-certs]
+version = "0.6"
 optional = true
 
 [dev-dependencies.uuid]
@@ -174,15 +169,15 @@ all = [
     "integrated-auth-gssapi",
     "rust_decimal",
     "bigdecimal",
-    "use_rustls",
+    "native-tls",
 ]
-default = ["tds73", "winauth"]
-vendored-openssl = ["opentls/vendored", "async-native-tls/vendored"]
+default = ["tds73", "winauth", "native-tls"]
 tds73 = []
 docs = []
 sql-browser-async-std = ["async-std"]
 sql-browser-tokio = ["tokio", "tokio-util"]
 sql-browser-smol = ["async-io", "async-net", "futures-lite"]
 integrated-auth-gssapi = ["libgssapi"]
-bigdecimal = [ "bigdecimal_"]
-use_rustls = ["rustls", "tokio-rustls", "tokio-util", "hyper-rustls", "rustls-pemfile"]
+bigdecimal = ["bigdecimal_"]
+rustls = ["tokio-rustls", "tokio-util", "rustls-pemfile", "rustls-native-certs"]
+native-tls = ["async-native-tls"]
