+++ src/error.rs
@@ -68,22 +68,7 @@ impl From<uuid::Error> for Error {
     }
 }
 
-#[cfg(any(target_os = "macos", target_os = "ios"))]
-#[cfg_attr(
-    feature = "docs",
-    doc(cfg(any(target_os = "macos", target_os = "ios")))
-)]
-impl From<opentls::Error> for Error {
-    fn from(v: opentls::Error) -> Self {
-        Error::Tls(format!("{}", v))
-    }
-}
-
-#[cfg(not(any(target_os = "macos", target_os = "ios")))]
-#[cfg_attr(
-    feature = "docs",
-    doc(cfg(not(any(target_os = "macos", target_os = "ios"))))
-)]
+#[cfg(feature = "native-tls")]
 impl From<async_native_tls::Error> for Error {
     fn from(v: async_native_tls::Error) -> Self {
         Error::Tls(format!("{}", v))
